Edited js structure
In this commit, the .js structure was changed to modules for easier workflow.
